refactor: cli now in one npm script and updated documtation
