+++ src/cli/apps/apps.ts
@@ -1,33 +1,30 @@
 import { PublicKey } from '@greymass/eosio';
-import { setSettings } from './src/settings';
-import { App, AppCreateOptions } from './src/app';
+import { setSettings } from '../../sdk/settings';
+import { App, AppCreateOptions } from '../../sdk/app';
 
-const args: string[] = process.argv.slice(2);
+export default async function apps(args: string[]) {
+    if (args[0] === 'create') {
+        console.log('Creating new app');
 
-async function main() {
-    console.log('Creating new app');
-    console.log('Arguments: ', args);
+        const options: AppCreateOptions = {
+            appName: args[0],
+            usernamePrefix: args[1],
+            description: args[2],
+            logoUrl: args[3],
+            origin: args[4],
+            publicKey: PublicKey.from(args[5]),
+        };
+        const blockchainUrl = args[6];
 
-    const options: AppCreateOptions = {
-        appName: args[0],
-        usernamePrefix: args[1],
-        description: args[2],
-        logoUrl: args[3],
-        origin: args[4],
-        publicKey: PublicKey.from(args[5]),
-    };
-    const blockchainUrl = args[6];
-    setSettings({
-        blockchainUrl,
-    });
+        setSettings({
+            blockchainUrl,
+        });
 
-    const res = await App.create(options);
+        const res = await App.create(options);
 
-    console.log('New app created with username: ', res.username.username);
-    console.log('and account name: ', res.accountName.toString());
+        console.log('New app created with username: ', res.username.username);
+        console.log('and account name: ', res.accountName.toString());
+    } else {
+        throw new Error(`Unknown command ${args[0]}`);
+    }
 }
-
-Promise.resolve(main()).catch((err) => {
-    console.error(err);
-    process.exit(1);
-});
